code typo in Getting_Started.md
Changed 
`Variable Declarions` should be near the top of the file
to 
`Variable Declarations` should be near the top of the file

I'm pretty new to all this, so I apologize if there's anything uncool about what/how I did.
Thanks for all the mods, and EXTENSIVE documention.  Very educational!