discord link in side bar, fixed search results being under version. ads
Look, there have always been ads, I used to have them by using a nginx subfilter on to inject them to the bottom of the page, it used to look like this https://i.blamejared.com/lliPL.png, but with this theme change they will be part of the theme themselves.

If the theme is ever published the codes will be moved to a theme config, but for now this is how it will be.