feat(mods): add Botany Pots documentation
I took note of a difference in the 1.15 version of the mod compared to its 1.14 version. Hopefully that is clear enough for users.

As well, in a very few instances, sub-headers were added directly after a header (h2 -> h3 -> text) and the generated code from the Docs-Site causes both headings to show on a single line. This has been reported: https://github.com/CraftTweaker/Docs-Site/issues/1