feat(1.14/mods): add Botany Pots documentation (#305)
* feat(mods): add Botany Pots documentation

I took note of a difference in the 1.15 version of the mod compared to its 1.14 version. Hopefully that is clear enough for users.

As well, in a very few instances, sub-headers were added directly after a header (h2 -> h3 -> text) and the generated code from the Docs-Site causes both headings to show on a single line. This has been reported: https://github.com/CraftTweaker/Docs-Site/issues/1

* fix(mods/botanypots): parameter links to raw file

* feat(mods/botanypots): add return type for `getAllIds()`

* feat(mods/botanypots): add example of logging ids to log file

* refactor(mods/botanypots): remove `x.getAllIds();` from code block

I had intended to keep this, but since the "base example" is already included in the inline code at the start of the sections, I'm opting to remove it to just leave an actual usage example in the code block.

* refactor: remove references of multiplier not existing in 1.15

Since the documentation is now split between the two versions, the mentions of the multiplier not existing no longer needs to exist.